Fix for issue #4.
Reorganized the language array to make it more readable.

Refactored the language detection code to set the default language if one is not already set in the settings.php file.  Also added logic so that the app can detect the user's browser preferences and display the site accordingly.

Created a method that returns a language drop down list.  This is better than having multiple static lists in various places.

Made the includes for the various language files more solid.  It checks to see if the file exists before trying to include it.  This should intercept any errors with the include.