feat(Password): Refactored the handling of resetting password
Changed the reset password screen so it no longr indicates if a
username/email was found.  Now it just displays a generic "If the user
you specified exists in our system, we will send a password to their
email address." message and silently fails if no user is
found.

Changes:
Moved the logic to the ResetPassword class

Added new
generic message to the lang_en file and updated the messages.php to
return only the generic message instead of a "success" message.

Removed additional tests since that functionality is no longer needed