New Feature: Email reset password link
Previously, when a user clicked the "forgot password" link, it would email their password to them in plaintext.  This was a horrible idea.

Today, when a user clicks the link, they are emailed a link where they can reset their password.  This functionality uses a "token" to confrim identity during the reset password process.

Commits:
    refactor(Messages): Moved "msg" to FlashMessage bag

    feat(Password): Added password reset link
      Instead of emailing the password to a user, an email with a "reset password" link is sent.
        - Updated corresponding files
        - Added email templates:
          - reset_password_link.html
          - forgot_password_success.html

    feat(Strings): Added new strings for forgot password
      - Added "email_forgot_pwd_subject"
      - Added "password_successful_changed"
      - Added "password_reset_confirmation_subject"
      - Added "password_confirm_blank"

    feat(Exceptions): Adding new exceptions
      - Added SendNotificationFailed
      - TokenExpired
      - TokenGenerationFailed