fix(Settings): Fix Warning: Illegal string offset error
The issue was the $langSelected was a string instead of an array thus
the error when trying to use the "string" as an array.  Changed it so
the default is an array with a default value of "en".

Also changed the
verbiage in updatesettings.php from "blank" to "not selected"

Closes #57