feat(Auth): Added "auth" attribute to session
Added "auth" boolean flag to the session that gets set on
login/logout.

Based on the flag value, if a user tries to go to a
non-view page and gets redirected to the index.php it will check to see
if they are already authenticated or not and redirect to the login view
or the home view.

Added logic to the main index.php to check for the
auth flag for redirection